requirements.txt
#pip install -r requirements.txt
pydantic-ai
python-dotenv
asyncio
fastapi 
uvicorn
openai
jinja2
python-multipart